requirements.txt
Django==5.1.8
Pillow==11.1.0
